GRPO_Project_Iteration/BlackJack/requirements.txt
# For streamlit
tqdm=4.67.1
pytorch=2.7.0
numpy=2.2.6
streamlit=1.45.1
gymnasium[toy_text]>=0.29.1